♻️ refactor(paiir): narrow package export surface
## Summary

This commit narrows the public init export surface of paibox.paiir and its pipeline packages so frontend-facing imports stay small and backend/internal symbols move back to their owning modules.

## Changes

- reduce top-level paibox.paiir exports to the intended frontend-facing API
- trim re-exports from paibox.paiir.pipeline and paibox.paiir.pipeline.avgpool
- update direct import sites in backend/docs/tests to import from the narrower owning modules
- refresh backend guide examples to match the new package boundaries

## Impact

- internal IR and backend-oriented symbols are no longer implicitly exposed from broad package roots
- call sites now depend on narrower, more explicit module ownership
- the public package surface is easier to understand and maintain
