+++ paibox/paiir/pipeline/__init__.py
@@ -1,26 +1,8 @@
 """PAIIR compile-time pipeline utilities."""
 
 from .compile import CompileConfig, compile_to_paiir
-from .data_format import (
-    DataFormat,
-    infer_output_format,
-    infer_weight_format,
-    merge_data_formats,
-)
-from .passes import (
-    propagate_signal_domain,
-    specialize_general_adds,
-    validate_deployable_graph,
-)
 
 __all__ = [
     "CompileConfig",
-    "DataFormat",
     "compile_to_paiir",
-    "infer_output_format",
-    "infer_weight_format",
-    "merge_data_formats",
-    "propagate_signal_domain",
-    "specialize_general_adds",
-    "validate_deployable_graph",
 ]
